+++ inference_graph/pipeline.config
@@ -1,0 +1,130 @@
+model {
+  faster_rcnn {
+    num_classes: 1
+    image_resizer {
+      keep_aspect_ratio_resizer {
+        min_dimension: 600
+        max_dimension: 1024
+      }
+    }
+    feature_extractor {
+      type: "faster_rcnn_inception_v2"
+      first_stage_features_stride: 16
+    }
+    first_stage_anchor_generator {
+      grid_anchor_generator {
+        height_stride: 16
+        width_stride: 16
+        scales: 0.25
+        scales: 0.5
+        scales: 1.0
+        scales: 2.0
+        aspect_ratios: 0.5
+        aspect_ratios: 1.0
+        aspect_ratios: 2.0
+      }
+    }
+    first_stage_box_predictor_conv_hyperparams {
+      op: CONV
+      regularizer {
+        l2_regularizer {
+          weight: 0.0
+        }
+      }
+      initializer {
+        truncated_normal_initializer {
+          stddev: 0.009999999776482582
+        }
+      }
+    }
+    first_stage_nms_score_threshold: 0.0
+    first_stage_nms_iou_threshold: 0.699999988079071
+    first_stage_max_proposals: 100
+    first_stage_localization_loss_weight: 2.0
+    first_stage_objectness_loss_weight: 1.0
+    initial_crop_size: 14
+    maxpool_kernel_size: 2
+    maxpool_stride: 2
+    second_stage_box_predictor {
+      mask_rcnn_box_predictor {
+        fc_hyperparams {
+          op: FC
+          regularizer {
+            l2_regularizer {
+              weight: 0.0
+            }
+          }
+          initializer {
+            variance_scaling_initializer {
+              factor: 1.0
+              uniform: true
+              mode: FAN_AVG
+            }
+          }
+        }
+        use_dropout: false
+        dropout_keep_probability: 1.0
+      }
+    }
+    second_stage_post_processing {
+      batch_non_max_suppression {
+        score_threshold: 0.30000001192092896
+        iou_threshold: 0.6000000238418579
+        max_detections_per_class: 100
+        max_total_detections: 100
+      }
+      score_converter: SOFTMAX
+    }
+    second_stage_localization_loss_weight: 2.0
+    second_stage_classification_loss_weight: 1.0
+  }
+}
+train_config {
+  batch_size: 1
+  data_augmentation_options {
+    random_horizontal_flip {
+    }
+  }
+  optimizer {
+    momentum_optimizer {
+      learning_rate {
+        manual_step_learning_rate {
+          initial_learning_rate: 0.00019999999494757503
+          schedule {
+            step: 900000
+            learning_rate: 1.9999999494757503e-05
+          }
+          schedule {
+            step: 1200000
+            learning_rate: 1.9999999949504854e-06
+          }
+        }
+      }
+      momentum_optimizer_value: 0.8999999761581421
+    }
+    use_moving_average: false
+  }
+  gradient_clipping_by_norm: 10.0
+  fine_tune_checkpoint: "C:/Users/Kinoshita_Takafumi/Documents/Tensorflow/models/research/object_detection/faster_rcnn_inception_v2_coco_2018_01_28/model.ckpt"
+  from_detection_checkpoint: true
+  num_steps: 50000
+}
+train_input_reader {
+  label_map_path: "C:/Users/Kinoshita_Takafumi/Documents/Tensorflow/models/research/object_detection/training/fddb_labelmap.pbtxt"
+  tf_record_input_reader {
+    input_path: "C:/Users/Kinoshita_Takafumi/Documents/Tensorflow/models/research/object_detection/fddb_train.record"
+  }
+}
+eval_config {
+  num_examples: 93
+  max_evals: 10
+  use_moving_averages: false
+}
+eval_input_reader {
+  label_map_path: "C:/Users/Kinoshita_Takafumi/Documents/Tensorflow/models/research/object_detection/training/fddb_labelmap.pbtxt"
+  shuffle: false
+  num_readers: 1
+  tf_record_input_reader {
+    input_path: "C:/Users/Kinoshita_Takafumi/Documents/Tensorflow/models/research/object_detection/fddb_val.record"
+  }
+}
