+++ training/fddb_labelmap.pbtxt
@@ -1,0 +1,4 @@
+item {
+    name: "face"
+    id: 1
+}
